6610 refactor and move search related discovery files
