--- Core/src/org/sleuthkit/autopsy/discovery/Bundle.properties
@@ -1,65 +1,0 @@
-FileSearchDialog.jLabel1.text=File Type
-FileSearchDialog.dsCheckBox.text=Data source
-FileSearchDialog.cancelButton.text=Cancel
-FileSearchDialog.freqCheckBox.text=CR Frequency
-FileSearchDialog.sizeCheckBox.text=Size
-FileSearchDialog.kwCheckBox.text=Keyword
-FileSearchDialog.addParentButton.text=Add
-FileSearchDialog.deleteParentButton.text=Delete
-FileSearchDialog.parentFullRadioButton.text=Full
-FileSearchDialog.parentSubstringRadioButton.text=Substring
-FileSearchDialog.jLabel2.text=(All will be used)
-FileSearchDialog.jLabel3.text=Group by attribute:
-FileSearchDialog.jLabel4.text=Order groups by:
-FileSearchDialog.orderAttrRadioButton.text=Attribute
-FileSearchDialog.orderSizeRadioButton.text=Group Size
-FileSearchDialog.jLabel5.text=Order files by: 
-FileSearchDialog.parentCheckBox.text=Parent
-FileSearchPanel.sortingPanel.border.title=Grouping
-FileSearchPanel.addButton.text=Add
-FileSearchPanel.substringRadioButton.text=Substring
-FileSearchPanel.fullRadioButton.text=Full
-FileSearchPanel.parentCheckbox.text=Parent Folder:
-FileSearchPanel.keywordCheckbox.text=Keyword:
-FileSearchPanel.crFrequencyCheckbox.text=Past Occurrences:
-FileSearchPanel.dataSourceCheckbox.text=Data Source:
-FileSearchPanel.sizeCheckbox.text=File Size:
-FileSearchPanel.orderGroupsByLabel.text=Order Groups By:
-FileSearchPanel.filtersScrollPane.border.title=Filters
-FileSearchPanel.parentLabel.text=(All will be used)
-FileSearchPanel.deleteButton.text=Delete
-FileSearchPanel.orderByLabel.text=Order Within Groups By:
-FileSearchPanel.groupByLabel.text=Group By:
-FileSearchDialog.searchButton.text=Search
-FileSearchDialog.hashCheckBox.text=Hash Set
-FileSearchDialog.intCheckBox.text=Interesting Items
-FileSearchDialog.tagsCheckBox.text=Tags
-FileSearchDialog.objCheckBox.text=Objects
-FileSearchDialog.exifCheckBox.text=Must contain EXIF data
-FileSearchDialog.notableCheckBox.text=Must have been tagged as notable
-FileSearchDialog.scoreCheckBox.text=Has score
-FileSearchPanel.hashSetCheckbox.text=Hash Set:
-FileSearchPanel.tagsCheckbox.text=Tag:
-FileSearchPanel.interestingItemsCheckbox.text=Interesting Item:
-FileSearchPanel.scoreCheckbox.text=Has Score:
-FileSearchPanel.notableCheckbox.text=Must have been tagged as notable
-FileSearchPanel.objectsCheckbox.text=Object Detected:
-DiscoveryExtractAction.title.extractFiles.text=Extract File
-FileSearchPanel.includeRadioButton.text=Include
-FileSearchPanel.excludeRadioButton.text=Exclude
-FileSearchPanel.knownFilesCheckbox.toolTipText=
-FileSearchPanel.knownFilesCheckbox.text=Hide known files
-FileSearchPanel.stepThreeLabel.text=Step 3: Choose display settings
-FileSearchPanel.userCreatedCheckbox.text=Possibly User Created
-UserCreatedFilterPanel.userCreatedCheckbox.text=Possibly User Created
-InterestingItemFilterPanel.interestingItemsCheckbox.text=Interesting Item:
-ParentFolderFilterPanel.parentCheckbox.text=Parent Folder:
-ParentFolderFilterPanel.deleteButton.text=Delete
-ParentFolderFilterPanel.excludeRadioButton.text=Exclude
-ParentFolderFilterPanel.includeRadioButton.text=Include
-ParentFolderFilterPanel.substringRadioButton.text=Substring
-ParentFolderFilterPanel.fullRadioButton.text=Full
-ParentFolderFilterPanel.parentLabel.text=(All will be used)
-ParentFolderFilterPanel.addButton.text=Add
-ObjectDetectedFilterPanel.text=Object Detected:
-DomainUniquenessFilterPanel.domainUniquenessCheckbox.text=Domain Uniqueness:
