+++ Core/src/org/sleuthkit/autopsy/discovery/search/FileSearch.java
@@ -16,27 +16,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.sleuthkit.autopsy.discovery;
+package org.sleuthkit.autopsy.discovery.search;
 
-import org.sleuthkit.autopsy.discovery.ui.VideoThumbnailsWrapper;
-import org.sleuthkit.autopsy.discovery.ui.AbstractFilter;
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
-import com.google.common.io.Files;
-import java.awt.Image;
-import java.awt.image.BufferedImage;
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.io.Reader;
-import java.nio.file.Paths;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -47,30 +35,15 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.logging.Level;
-import javax.imageio.ImageIO;
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang.StringUtils;
-import org.imgscalr.Scalr;
-import org.netbeans.api.progress.ProgressHandle;
-import org.opencv.core.Mat;
-import org.opencv.highgui.VideoCapture;
-import org.openide.util.Lookup;
 import org.openide.util.NbBundle;
-import org.sleuthkit.autopsy.casemodule.Case;
-import org.sleuthkit.autopsy.casemodule.NoCurrentCaseException;
 import org.sleuthkit.autopsy.centralrepository.datamodel.CorrelationAttributeInstance;
 import org.sleuthkit.autopsy.centralrepository.datamodel.CentralRepoException;
 import org.sleuthkit.autopsy.centralrepository.datamodel.CentralRepoDbUtil;
 import org.sleuthkit.autopsy.centralrepository.datamodel.InstanceTableCallback;
-import org.sleuthkit.autopsy.corelibs.ScalrWrapper;
-import org.sleuthkit.autopsy.coreutils.ImageUtils;
 import org.sleuthkit.autopsy.coreutils.Logger;
-import static org.sleuthkit.autopsy.coreutils.VideoUtils.getVideoFileInTempDir;
-import org.sleuthkit.autopsy.datamodel.ContentUtils;
-import org.sleuthkit.autopsy.discovery.FileSearchData.FileSize;
-import org.sleuthkit.autopsy.discovery.FileSearchData.FileType;
-import org.sleuthkit.autopsy.discovery.FileSearchData.Frequency;
+import org.sleuthkit.autopsy.discovery.search.FileSearchData.FileType;
+import org.sleuthkit.autopsy.discovery.search.FileSearchData.Frequency;
 import org.sleuthkit.datamodel.AbstractFile;
 import org.sleuthkit.datamodel.BlackboardArtifact;
 import org.sleuthkit.datamodel.BlackboardAttribute;
@@ -81,29 +54,22 @@ import org.sleuthkit.datamodel.SleuthkitCase;
 import org.sleuthkit.datamodel.TskCoreException;
 import org.sleuthkit.datamodel.TskData;
 import org.sleuthkit.autopsy.centralrepository.datamodel.CentralRepository;
-import org.sleuthkit.autopsy.textextractors.TextExtractor;
-import org.sleuthkit.autopsy.textextractors.TextExtractorFactory;
+import org.sleuthkit.autopsy.discovery.search.DiscoveryKeyUtils.GroupKey;
+import org.sleuthkit.autopsy.discovery.search.DiscoveryKeyUtils.SearchKey;
 import org.sleuthkit.autopsy.textsummarizer.TextSummarizer;
 import org.sleuthkit.autopsy.textsummarizer.TextSummary;
-import org.sleuthkit.autopsy.texttranslation.NoServiceProviderException;
-import org.sleuthkit.autopsy.texttranslation.TextTranslationService;
-import org.sleuthkit.autopsy.texttranslation.TranslationException;
 
 /**
  * Main class to perform the file search.
  */
-class FileSearch {
+public class FileSearch {
 
     private final static Logger logger = Logger.getLogger(FileSearch.class.getName());
     private static final int MAXIMUM_CACHE_SIZE = 10;
-    private static final String THUMBNAIL_FORMAT = "png"; //NON-NLS
-    private static final String VIDEO_THUMBNAIL_DIR = "video-thumbnails"; //NON-NLS
     private static final Cache<SearchKey, Map<GroupKey, List<ResultFile>>> searchCache = CacheBuilder.newBuilder()
             .maximumSize(MAXIMUM_CACHE_SIZE)
             .build();
-    private static final int PREVIEW_SIZE = 256;
-    private static volatile TextSummarizer summarizerToUse = null;
-    private static final BufferedImage VIDEO_DEFAULT_IMAGE = getDefaultVideoThumbnail();
+
 
     /**
      * Run the file search and returns the SearchResults object for debugging.
@@ -175,7 +141,7 @@ class FileSearch {
      *
      * @throws FileSearchException
      */
-    static Map<GroupKey, Integer> getGroupSizes(String userName,
+    public static Map<GroupKey, Integer> getGroupSizes(String userName,
             List<AbstractFilter> filters,
             AttributeType groupAttributeType,
             FileGroup.GroupSortingAlgorithm groupSortingType,
@@ -212,7 +178,7 @@ class FileSearch {
      *
      * @throws FileSearchException
      */
-    static List<ResultFile> getFilesInGroup(String userName,
+    public static List<ResultFile> getFilesInGroup(String userName,
             List<AbstractFilter> filters,
             AttributeType groupAttributeType,
             FileGroup.GroupSortingAlgorithm groupSortingType,
@@ -269,15 +235,12 @@ class FileSearch {
      */
     @NbBundle.Messages({"FileSearch.documentSummary.noPreview=No preview available.",
         "FileSearch.documentSummary.noBytes=No bytes read for document, unable to display preview."})
-    static TextSummary summarize(AbstractFile file) {
+    public static TextSummary summarize(AbstractFile file) {
         TextSummary summary = null;
-        TextSummarizer localSummarizer = summarizerToUse;
-        if (localSummarizer == null) {
-            synchronized (searchCache) {
-                if (localSummarizer == null) {
-                    localSummarizer = getLocalSummarizer();
-                }
-            }
+        TextSummarizer localSummarizer; 
+        synchronized (searchCache) {
+            localSummarizer = SummaryHelpers.getLocalSummarizer();
+
         }
         if (localSummarizer != null) {
             try {
@@ -289,182 +252,9 @@ class FileSearch {
         }
         if (summary == null || StringUtils.isBlank(summary.getSummaryText())) {
             //summary text was empty grab the beginning of the file 
-            summary = getDefaultSummary(file);
+            summary = SummaryHelpers.getDefaultSummary(file);
         }
         return summary;
-    }
-
-    private static TextSummary getDefaultSummary(AbstractFile file) {
-        Image image = null;
-        int countOfImages = 0;
-        try {
-            Content largestChild = null;
-            for (Content child : file.getChildren()) {
-                if (child instanceof AbstractFile && ImageUtils.isImageThumbnailSupported((AbstractFile) child)) {
-                    countOfImages++;
-                    if (largestChild == null || child.getSize() > largestChild.getSize()) {
-                        largestChild = child;
-                    }
-                }
-            }
-            if (largestChild != null) {
-                image = ImageUtils.getThumbnail(largestChild, ImageUtils.ICON_SIZE_LARGE);
-            }
-        } catch (TskCoreException ex) {
-            logger.log(Level.WARNING, "Error getting children for file: " + file.getId(), ex);
-        }
-        image = image == null ? image : image.getScaledInstance(ImageUtils.ICON_SIZE_MEDIUM, ImageUtils.ICON_SIZE_MEDIUM,
-                Image.SCALE_SMOOTH);
-        String summaryText = null;
-        if (file.getMd5Hash() != null) {
-            try {
-                summaryText = getSavedSummary(Paths.get(Case.getCurrentCaseThrows().getCacheDirectory(), "summaries", file.getMd5Hash() + "-default-" + PREVIEW_SIZE + "-translated.txt").toString());
-            } catch (NoCurrentCaseException ex) {
-                logger.log(Level.WARNING, "Unable to retrieve saved summary. No case is open.", ex);
-            }
-        }
-        if (StringUtils.isBlank(summaryText)) {
-            String firstLines = getFirstLines(file);
-            String translatedFirstLines = getTranslatedVersion(firstLines);
-            if (!StringUtils.isBlank(translatedFirstLines)) {
-                summaryText = translatedFirstLines;
-                if (file.getMd5Hash() != null) {
-                    try {
-                        saveSummary(summaryText, Paths.get(Case.getCurrentCaseThrows().getCacheDirectory(), "summaries", file.getMd5Hash() + "-default-" + PREVIEW_SIZE + "-translated.txt").toString());
-                    } catch (NoCurrentCaseException ex) {
-                        logger.log(Level.WARNING, "Unable to save translated summary. No case is open.", ex);
-                    }
-                }
-            } else {
-                summaryText = firstLines;
-            }
-        }
-        return new TextSummary(summaryText, image, countOfImages);
-    }
-
-    /**
-     * Provide an English version of the specified String if it is not English,
-     * translation is enabled, and it can be translated.
-     *
-     * @param documentString The String to provide an English version of.
-     *
-     * @return The English version of the provided String, or null if no
-     *         translation occurred.
-     */
-    private static String getTranslatedVersion(String documentString) {
-        try {
-            TextTranslationService translatorInstance = TextTranslationService.getInstance();
-            if (translatorInstance.hasProvider()) {
-                String translatedResult = translatorInstance.translate(documentString);
-                if (translatedResult.isEmpty() == false) {
-                    return translatedResult;
-                }
-            }
-        } catch (NoServiceProviderException | TranslationException ex) {
-            logger.log(Level.INFO, "Error translating string for summary", ex);
-        }
-        return null;
-    }
-
-    /**
-     * Find and load a saved summary from the case folder for the specified
-     * file.
-     *
-     * @param summarySavePath The full path for the saved summary file.
-     *
-     * @return The summary found given the specified path, null if no summary
-     *         was found.
-     */
-    private static String getSavedSummary(String summarySavePath) {
-        if (summarySavePath == null) {
-            return null;
-        }
-        File savedFile = new File(summarySavePath);
-        if (savedFile.exists()) {
-            try (BufferedReader bReader = new BufferedReader(new FileReader(savedFile))) {
-                // pass the path to the file as a parameter
-                StringBuilder sBuilder = new StringBuilder();
-                String sCurrentLine = bReader.readLine();
-                while (sCurrentLine != null) {
-                    sBuilder.append(sCurrentLine).append('\n');
-                    sCurrentLine = bReader.readLine();
-                }
-                return sBuilder.toString();
-            } catch (IOException ingored) {
-                //summary file may not exist or may be incomplete in which case return null so a summary can be generated
-                return null; //no saved summary was able to be found
-            }
-        } else {
-            try {  //if the file didn't exist make sure the parent directories exist before we move on to creating a summary
-                Files.createParentDirs(savedFile);
-            } catch (IOException ex) {
-                logger.log(Level.WARNING, "Unable to create summaries directory in case folder for file at: " + summarySavePath, ex);
-            }
-            return null; //no saved summary was able to be found
-        }
-
-    }
-
-    /**
-     * Save a summary at the specified location.
-     *
-     * @param summary         The text of the summary being saved.
-     * @param summarySavePath The full path for the saved summary file.
-     */
-    private static void saveSummary(String summary, String summarySavePath) {
-        if (summarySavePath == null) {
-            return;  //can't save a summary if we don't have a path
-        }
-        try (FileWriter myWriter = new FileWriter(summarySavePath)) {
-            myWriter.write(summary);
-        } catch (IOException ex) {
-            logger.log(Level.WARNING, "Unable to save summary at: " + summarySavePath, ex);
-        }
-    }
-
-    /**
-     * Get the beginning of text from the specified AbstractFile.
-     *
-     * @param file The AbstractFile to get text from.
-     *
-     * @return The beginning of text from the specified AbstractFile.
-     */
-    private static String getFirstLines(AbstractFile file) {
-        TextExtractor extractor;
-        try {
-            extractor = TextExtractorFactory.getExtractor(file, null);
-        } catch (TextExtractorFactory.NoTextExtractorFound ignored) {
-            //no extractor found, use Strings Extractor
-            extractor = TextExtractorFactory.getStringsExtractor(file, null);
-        }
-
-        try (Reader reader = extractor.getReader()) {
-            char[] cbuf = new char[PREVIEW_SIZE];
-            reader.read(cbuf, 0, PREVIEW_SIZE);
-            return new String(cbuf);
-        } catch (IOException ex) {
-            return Bundle.FileSearch_documentSummary_noBytes();
-        } catch (TextExtractor.InitReaderException ex) {
-            return Bundle.FileSearch_documentSummary_noPreview();
-        }
-    }
-
-    /**
-     * Get the first TextSummarizer found by a lookup of TextSummarizers.
-     *
-     * @return The first TextSummarizer found by a lookup of TextSummarizers.
-     *
-     * @throws IOException
-     */
-    private static TextSummarizer getLocalSummarizer() {
-        Collection<? extends TextSummarizer> summarizers
-                = Lookup.getDefault().lookupAll(TextSummarizer.class
-                );
-        if (!summarizers.isEmpty()) {
-            summarizerToUse = summarizers.iterator().next();
-            return summarizerToUse;
-        }
-        return null;
     }
 
     /**
@@ -593,271 +383,6 @@ class FileSearch {
                 + "AND blackboard_artifacts.obj_id IN (" + objIdList + ") "; // NON-NLS
     }
 
-    /**
-     * Get the default image to display when a thumbnail is not available.
-     *
-     * @return The default video thumbnail.
-     */
-    private static BufferedImage getDefaultVideoThumbnail() {
-        try {
-            return ImageIO.read(ImageUtils.class.getResourceAsStream("/org/sleuthkit/autopsy/images/failedToCreateVideoThumb.png"));//NON-NLS
-        } catch (IOException ex) {
-            logger.log(Level.SEVERE, "Failed to load 'failed to create video' placeholder.", ex); //NON-NLS
-        }
-        return null;
-    }
-
-    /**
-     * Get the video thumbnails for a file which exists in a
-     * VideoThumbnailsWrapper and update the VideoThumbnailsWrapper to include
-     * them.
-     *
-     * @param thumbnailWrapper the object which contains the file to generate
-     *                         thumbnails for.
-     *
-     */
-    @NbBundle.Messages({"# {0} - file name",
-        "FileSearch.genVideoThumb.progress.text=extracting temporary file {0}"})
-    static void getVideoThumbnails(VideoThumbnailsWrapper thumbnailWrapper) {
-        AbstractFile file = thumbnailWrapper.getResultFile().getFirstInstance();
-        String cacheDirectory;
-        try {
-            cacheDirectory = Case.getCurrentCaseThrows().getCacheDirectory();
-        } catch (NoCurrentCaseException ex) {
-            cacheDirectory = null;
-            logger.log(Level.WARNING, "Unable to get cache directory, video thumbnails will not be saved", ex);
-        }
-        if (cacheDirectory == null || file.getMd5Hash() == null || !Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash()).toFile().exists()) {
-            java.io.File tempFile;
-            try {
-                tempFile = getVideoFileInTempDir(file);
-            } catch (NoCurrentCaseException ex) {
-                logger.log(Level.WARNING, "Exception while getting open case.", ex); //NON-NLS
-                int[] framePositions = new int[]{
-                    0,
-                    0,
-                    0,
-                    0};
-                thumbnailWrapper.setThumbnails(createDefaultThumbnailList(VIDEO_DEFAULT_IMAGE), framePositions);
-                return;
-            }
-            if (tempFile.exists() == false || tempFile.length() < file.getSize()) {
-                ProgressHandle progress = ProgressHandle.createHandle(Bundle.FileSearch_genVideoThumb_progress_text(file.getName()));
-                progress.start(100);
-                try {
-                    Files.createParentDirs(tempFile);
-                    if (Thread.interrupted()) {
-                        int[] framePositions = new int[]{
-                            0,
-                            0,
-                            0,
-                            0};
-                        thumbnailWrapper.setThumbnails(createDefaultThumbnailList(VIDEO_DEFAULT_IMAGE), framePositions);
-                        return;
-                    }
-                    ContentUtils.writeToFile(file, tempFile, progress, null, true);
-                } catch (IOException ex) {
-                    logger.log(Level.WARNING, "Error extracting temporary file for " + file.getParentPath() + "/" + file.getName(), ex); //NON-NLS
-                } finally {
-                    progress.finish();
-                }
-            }
-            VideoCapture videoFile = new VideoCapture(); // will contain the video
-            BufferedImage bufferedImage = null;
-
-            try {
-                if (!videoFile.open(tempFile.toString())) {
-                    logger.log(Level.WARNING, "Error opening {0} for preview generation.", file.getParentPath() + "/" + file.getName()); //NON-NLS
-                    int[] framePositions = new int[]{
-                        0,
-                        0,
-                        0,
-                        0};
-                    thumbnailWrapper.setThumbnails(createDefaultThumbnailList(VIDEO_DEFAULT_IMAGE), framePositions);
-                    return;
-                }
-                double fps = videoFile.get(5); // gets frame per second
-                double totalFrames = videoFile.get(7); // gets total frames
-                if (fps <= 0 || totalFrames <= 0) {
-                    logger.log(Level.WARNING, "Error getting fps or total frames for {0}", file.getParentPath() + "/" + file.getName()); //NON-NLS
-                    int[] framePositions = new int[]{
-                        0,
-                        0,
-                        0,
-                        0};
-                    thumbnailWrapper.setThumbnails(createDefaultThumbnailList(VIDEO_DEFAULT_IMAGE), framePositions);
-                    return;
-                }
-                if (Thread.interrupted()) {
-                    int[] framePositions = new int[]{
-                        0,
-                        0,
-                        0,
-                        0};
-                    thumbnailWrapper.setThumbnails(createDefaultThumbnailList(VIDEO_DEFAULT_IMAGE), framePositions);
-                    return;
-                }
-
-                double duration = 1000 * (totalFrames / fps); //total milliseconds
-
-                int[] framePositions = new int[]{
-                    (int) (duration * .01),
-                    (int) (duration * .25),
-                    (int) (duration * .5),
-                    (int) (duration * .75),};
-
-                Mat imageMatrix = new Mat();
-                List<Image> videoThumbnails = new ArrayList<>();
-                if (cacheDirectory == null || file.getMd5Hash() == null) {
-                    cacheDirectory = null;
-                } else {
-                    try {
-                        FileUtils.forceMkdir(Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash()).toFile());
-                    } catch (IOException ex) {
-                        cacheDirectory = null;
-                        logger.log(Level.WARNING, "Unable to make video thumbnails directory, thumbnails will not be saved", ex);
-                    }
-                }
-                for (int i = 0; i < framePositions.length; i++) {
-                    if (!videoFile.set(0, framePositions[i])) {
-                        logger.log(Level.WARNING, "Error seeking to " + framePositions[i] + "ms in {0}", file.getParentPath() + "/" + file.getName()); //NON-NLS
-                        // If we can't set the time, continue to the next frame position and try again.
-
-                        videoThumbnails.add(VIDEO_DEFAULT_IMAGE);
-                        if (cacheDirectory != null) {
-                            try {
-                                ImageIO.write(VIDEO_DEFAULT_IMAGE, THUMBNAIL_FORMAT,
-                                        Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash(), i + "-" + framePositions[i] + "." + THUMBNAIL_FORMAT).toFile()); //NON-NLS)
-                            } catch (IOException ex) {
-                                logger.log(Level.WARNING, "Unable to save default video thumbnail for " + file.getMd5Hash() + " at frame position " + framePositions[i], ex);
-                            }
-                        }
-                        continue;
-                    }
-                    // Read the frame into the image/matrix.
-                    if (!videoFile.read(imageMatrix)) {
-                        logger.log(Level.WARNING, "Error reading frame at " + framePositions[i] + "ms from {0}", file.getParentPath() + "/" + file.getName()); //NON-NLS
-                        // If the image is bad for some reason, continue to the next frame position and try again.
-                        videoThumbnails.add(VIDEO_DEFAULT_IMAGE);
-                        if (cacheDirectory != null) {
-                            try {
-                                ImageIO.write(VIDEO_DEFAULT_IMAGE, THUMBNAIL_FORMAT,
-                                        Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash(), i + "-" + framePositions[i] + "." + THUMBNAIL_FORMAT).toFile()); //NON-NLS)
-                            } catch (IOException ex) {
-                                logger.log(Level.WARNING, "Unable to save default video thumbnail for " + file.getMd5Hash() + " at frame position " + framePositions[i], ex);
-                            }
-                        }
-
-                        continue;
-                    }
-                    // If the image is empty, return since no buffered image can be created.
-                    if (imageMatrix.empty()) {
-                        videoThumbnails.add(VIDEO_DEFAULT_IMAGE);
-                        if (cacheDirectory != null) {
-                            try {
-                                ImageIO.write(VIDEO_DEFAULT_IMAGE, THUMBNAIL_FORMAT,
-                                        Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash(), i + "-" + framePositions[i] + "." + THUMBNAIL_FORMAT).toFile()); //NON-NLS)
-                            } catch (IOException ex) {
-                                logger.log(Level.WARNING, "Unable to save default video thumbnail for " + file.getMd5Hash() + " at frame position " + framePositions[i], ex);
-                            }
-                        }
-                        continue;
-                    }
-
-                    int matrixColumns = imageMatrix.cols();
-                    int matrixRows = imageMatrix.rows();
-
-                    // Convert the matrix that contains the frame to a buffered image.
-                    if (bufferedImage == null) {
-                        bufferedImage = new BufferedImage(matrixColumns, matrixRows, BufferedImage.TYPE_3BYTE_BGR);
-                    }
-
-                    byte[] data = new byte[matrixRows * matrixColumns * (int) (imageMatrix.elemSize())];
-                    imageMatrix.get(0, 0, data); //copy the image to data
-
-                    if (imageMatrix.channels() == 3) {
-                        for (int k = 0; k < data.length; k += 3) {
-                            byte temp = data[k];
-                            data[k] = data[k + 2];
-                            data[k + 2] = temp;
-                        }
-                    }
-
-                    bufferedImage.getRaster().setDataElements(0, 0, matrixColumns, matrixRows, data);
-                    if (Thread.interrupted()) {
-                        thumbnailWrapper.setThumbnails(videoThumbnails, framePositions);
-                        try {
-                            FileUtils.forceDelete(Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash()).toFile());
-                        } catch (IOException ex) {
-                            logger.log(Level.WARNING, "Unable to delete directory for cancelled video thumbnail process", ex);
-                        }
-                        return;
-                    }
-                    BufferedImage thumbnail = ScalrWrapper.resize(bufferedImage, Scalr.Method.SPEED, Scalr.Mode.FIT_TO_HEIGHT, ImageUtils.ICON_SIZE_LARGE, ImageUtils.ICON_SIZE_MEDIUM, Scalr.OP_ANTIALIAS);
-                    //We are height limited here so it can be wider than it can be tall.Scalr maintains the aspect ratio.
-                    videoThumbnails.add(thumbnail);
-                    if (cacheDirectory != null) {
-                        try {
-                            ImageIO.write(thumbnail, THUMBNAIL_FORMAT,
-                                    Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash(), i + "-" + framePositions[i] + "." + THUMBNAIL_FORMAT).toFile()); //NON-NLS)
-                        } catch (IOException ex) {
-                            logger.log(Level.WARNING, "Unable to save video thumbnail for " + file.getMd5Hash() + " at frame position " + framePositions[i], ex);
-                        }
-                    }
-                }
-                thumbnailWrapper.setThumbnails(videoThumbnails, framePositions);
-            } finally {
-                videoFile.release(); // close the file}
-            }
-        } else {
-            loadSavedThumbnails(cacheDirectory, thumbnailWrapper, VIDEO_DEFAULT_IMAGE);
-        }
-    }
-
-    /**
-     * Load the thumbnails that exist in the cache directory for the specified
-     * video file.
-     *
-     * @param cacheDirectory   The directory which exists for the video
-     *                         thumbnails.
-     * @param thumbnailWrapper The VideoThumbnailWrapper object which contains
-     *                         information about the file and the thumbnails
-     *                         associated with it.
-     */
-    private static void loadSavedThumbnails(String cacheDirectory, VideoThumbnailsWrapper thumbnailWrapper, BufferedImage failedVideoThumbImage) {
-        int[] framePositions = new int[4];
-        List<Image> videoThumbnails = new ArrayList<>();
-        int thumbnailNumber = 0;
-        String md5 = thumbnailWrapper.getResultFile().getFirstInstance().getMd5Hash();
-        for (String fileName : Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, md5).toFile().list()) {
-            try {
-                videoThumbnails.add(ImageIO.read(Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, md5, fileName).toFile()));
-            } catch (IOException ex) {
-                videoThumbnails.add(failedVideoThumbImage);
-                logger.log(Level.WARNING, "Unable to read saved video thumbnail " + fileName + " for " + md5, ex);
-            }
-            int framePos = Integer.valueOf(FilenameUtils.getBaseName(fileName).substring(2));
-            framePositions[thumbnailNumber] = framePos;
-            thumbnailNumber++;
-        }
-        thumbnailWrapper.setThumbnails(videoThumbnails, framePositions);
-    }
-
-    /**
-     * Private helper method for creating video thumbnails, for use when no
-     * thumbnails are created.
-     *
-     * @return List containing the default thumbnail.
-     */
-    private static List<Image> createDefaultThumbnailList(BufferedImage failedVideoThumbImage) {
-        List<Image> videoThumbnails = new ArrayList<>();
-        videoThumbnails.add(failedVideoThumbImage);
-        videoThumbnails.add(failedVideoThumbImage);
-        videoThumbnails.add(failedVideoThumbImage);
-        videoThumbnails.add(failedVideoThumbImage);
-        return videoThumbnails;
-    }
-
     private FileSearch() {
         // Class should not be instantiated
     }
@@ -865,7 +390,7 @@ class FileSearch {
     /**
      * Base class for the grouping attributes.
      */
-    abstract static class AttributeType {
+    public abstract static class AttributeType {
 
         /**
          * For a given file, return the key for the group it belongs to for this
@@ -875,7 +400,7 @@ class FileSearch {
          *
          * @return the key for the group this file goes in
          */
-        abstract GroupKey getGroupKey(ResultFile file);
+        public abstract GroupKey getGroupKey(ResultFile file);
 
         /**
          * Add any extra data to the ResultFile object from this attribute.
@@ -887,60 +412,12 @@ class FileSearch {
          *
          * @throws FileSearchException
          */
-        void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb, CentralRepository centralRepoDb) throws FileSearchException {
+        public void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb, CentralRepository centralRepoDb) throws FileSearchException {
             // Default is to do nothing
         }
     }
 
-    /**
-     * The key used for grouping for each attribute type.
-     */
-    abstract static class GroupKey implements Comparable<GroupKey> {
-
-        /**
-         * Get the string version of the group key for display. Each display
-         * name should correspond to a unique GroupKey object.
-         *
-         * @return The display name for this key
-         */
-        abstract String getDisplayName();
-
-        /**
-         * Subclasses must implement equals().
-         *
-         * @param otherKey
-         *
-         * @return true if the keys are equal, false otherwise
-         */
-        @Override
-        abstract public boolean equals(Object otherKey);
-
-        /**
-         * Subclasses must implement hashCode().
-         *
-         * @return the hash code
-         */
-        @Override
-        abstract public int hashCode();
-
-        /**
-         * It should not happen with the current setup, but we need to cover the
-         * case where two different GroupKey subclasses are compared against
-         * each other. Use a lexicographic comparison on the class names.
-         *
-         * @param otherGroupKey The other group key
-         *
-         * @return result of alphabetical comparison on the class name
-         */
-        int compareClassNames(GroupKey otherGroupKey) {
-            return this.getClass().getName().compareTo(otherGroupKey.getClass().getName());
-        }
-
-        @Override
-        public String toString() {
-            return getDisplayName();
-        }
-    }
+   
 
     /**
      * Attribute for grouping/sorting by file size
@@ -948,75 +425,18 @@ class FileSearch {
     static class FileSizeAttribute extends AttributeType {
 
         @Override
-        GroupKey getGroupKey(ResultFile file) {
+        public GroupKey getGroupKey(ResultFile file) {
             return new FileSizeGroupKey(file);
-        }
-    }
-
-    /**
-     * Key representing a file size group
-     */
-    private static class FileSizeGroupKey extends GroupKey {
-
-        private final FileSize fileSize;
-
-        FileSizeGroupKey(ResultFile file) {
-            if (file.getFileType() == FileType.VIDEO) {
-                fileSize = FileSize.fromVideoSize(file.getFirstInstance().getSize());
-            } else {
-                fileSize = FileSize.fromImageSize(file.getFirstInstance().getSize());
-            }
-        }
-
-        @Override
-        String getDisplayName() {
-            return getFileSize().toString();
-        }
-
-        @Override
-        public int compareTo(GroupKey otherGroupKey) {
-            if (otherGroupKey instanceof FileSizeGroupKey) {
-                FileSizeGroupKey otherFileSizeGroupKey = (FileSizeGroupKey) otherGroupKey;
-                return Integer.compare(getFileSize().getRanking(), otherFileSizeGroupKey.getFileSize().getRanking());
-            } else {
-                return compareClassNames(otherGroupKey);
-            }
-        }
-
-        @Override
-        public boolean equals(Object otherKey) {
-            if (otherKey == this) {
-                return true;
-            }
-
-            if (!(otherKey instanceof FileSizeGroupKey)) {
-                return false;
-            }
-
-            FileSizeGroupKey otherFileSizeGroupKey = (FileSizeGroupKey) otherKey;
-            return getFileSize().equals(otherFileSizeGroupKey.getFileSize());
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(getFileSize().getRanking());
-        }
-
-        /**
-         * @return the fileSize
-         */
-        FileSize getFileSize() {
-            return fileSize;
         }
     }
 
     /**
      * Attribute for grouping/sorting by parent path
      */
-    static class ParentPathAttribute extends AttributeType {
+    public static class ParentPathAttribute extends AttributeType {
 
         @Override
-        GroupKey getGroupKey(ResultFile file) {
+        public GroupKey getGroupKey(ResultFile file) {
             return new ParentPathGroupKey(file);
         }
     }
@@ -1135,7 +555,7 @@ class FileSearch {
     static class DataSourceAttribute extends AttributeType {
 
         @Override
-        GroupKey getGroupKey(ResultFile file) {
+        public GroupKey getGroupKey(ResultFile file) {
             return new DataSourceGroupKey(file);
         }
     }
@@ -1215,7 +635,7 @@ class FileSearch {
     static class FileTypeAttribute extends AttributeType {
 
         @Override
-        GroupKey getGroupKey(ResultFile file) {
+        public GroupKey getGroupKey(ResultFile file) {
             return new FileTypeGroupKey(file);
         }
     }
@@ -1279,12 +699,12 @@ class FileSearch {
     static class KeywordListAttribute extends AttributeType {
 
         @Override
-        GroupKey getGroupKey(ResultFile file) {
+        public GroupKey getGroupKey(ResultFile file) {
             return new KeywordListGroupKey(file);
         }
 
         @Override
-        void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb,
+        public void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb,
                 CentralRepository centralRepoDb) throws FileSearchException {
 
             // Get pairs of (object ID, keyword list name) for all files in the list of files that have
@@ -1434,12 +854,12 @@ class FileSearch {
         static final int BATCH_SIZE = 50; // Number of hashes to look up at one time
 
         @Override
-        GroupKey getGroupKey(ResultFile file) {
+        public GroupKey getGroupKey(ResultFile file) {
             return new FrequencyGroupKey(file);
         }
 
         @Override
-        void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb,
+        public void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb,
                 CentralRepository centralRepoDb) throws FileSearchException {
             if (centralRepoDb == null) {
                 for (ResultFile file : files) {
@@ -1582,12 +1002,12 @@ class FileSearch {
     static class HashHitsAttribute extends AttributeType {
 
         @Override
-        GroupKey getGroupKey(ResultFile file) {
+        public GroupKey getGroupKey(ResultFile file) {
             return new HashHitsGroupKey(file);
         }
 
         @Override
-        void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb,
+        public void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb,
                 CentralRepository centralRepoDb) throws FileSearchException {
 
             // Get pairs of (object ID, hash set name) for all files in the list of files that have
@@ -1734,12 +1154,12 @@ class FileSearch {
     static class InterestingItemAttribute extends AttributeType {
 
         @Override
-        GroupKey getGroupKey(ResultFile file) {
+        public GroupKey getGroupKey(ResultFile file) {
             return new InterestingItemGroupKey(file);
         }
 
         @Override
-        void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb,
+        public void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb,
                 CentralRepository centralRepoDb) throws FileSearchException {
 
             // Get pairs of (object ID, interesting item set name) for all files in the list of files that have
@@ -1888,12 +1308,12 @@ class FileSearch {
     static class ObjectDetectedAttribute extends AttributeType {
 
         @Override
-        GroupKey getGroupKey(ResultFile file) {
+        public GroupKey getGroupKey(ResultFile file) {
             return new ObjectDetectedGroupKey(file);
         }
 
         @Override
-        void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb,
+        public void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb,
                 CentralRepository centralRepoDb) throws FileSearchException {
 
             // Get pairs of (object ID, object type name) for all files in the list of files that have
@@ -2041,12 +1461,12 @@ class FileSearch {
     static class FileTagAttribute extends AttributeType {
 
         @Override
-        GroupKey getGroupKey(ResultFile file) {
-            return new FileTagGroupKey(file);
+        public GroupKey getGroupKey(ResultFile file) {
+            return new DiscoveryKeyUtils.FileTagGroupKey(file);
         }
 
         @Override
-        void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb,
+        public void addAttributeToResultFiles(List<ResultFile> files, SleuthkitCase caseDb,
                 CentralRepository centralRepoDb) throws FileSearchException {
 
             try {
@@ -2063,202 +1483,7 @@ class FileSearch {
         }
     }
 
-    /**
-     * Represents a key for a specific search for a specific user.
-     */
-    private static class SearchKey implements Comparable<SearchKey> {
 
-        private final String keyString;
-
-        /**
-         * Construct a new SearchKey with all information that defines a search.
-         *
-         * @param userName           The name of the user performing the search.
-         * @param filters            The FileFilters being used for the search.
-         * @param groupAttributeType The AttributeType to group by.
-         * @param groupSortingType   The algorithm to sort the groups by.
-         * @param fileSortingMethod  The method to sort the files by.
-         */
-        SearchKey(String userName, List<AbstractFilter> filters,
-                AttributeType groupAttributeType,
-                FileGroup.GroupSortingAlgorithm groupSortingType,
-                FileSorter.SortingMethod fileSortingMethod) {
-            StringBuilder searchStringBuilder = new StringBuilder();
-            searchStringBuilder.append(userName);
-            for (AbstractFilter filter : filters) {
-                searchStringBuilder.append(filter.toString());
-            }
-            searchStringBuilder.append(groupAttributeType).append(groupSortingType).append(fileSortingMethod);
-            keyString = searchStringBuilder.toString();
-        }
-
-        @Override
-        public int compareTo(SearchKey otherSearchKey) {
-            return getKeyString().compareTo(otherSearchKey.getKeyString());
-        }
-
-        @Override
-        public boolean equals(Object otherKey) {
-            if (otherKey == this) {
-                return true;
-            }
-
-            if (!(otherKey instanceof SearchKey)) {
-                return false;
-            }
-
-            SearchKey otherSearchKey = (SearchKey) otherKey;
-            return getKeyString().equals(otherSearchKey.getKeyString());
-        }
-
-        @Override
-        public int hashCode() {
-            int hash = 5;
-            hash = 79 * hash + Objects.hashCode(getKeyString());
-            return hash;
-        }
-
-        /**
-         * @return the keyString
-         */
-        String getKeyString() {
-            return keyString;
-        }
-    }
-
-    /**
-     * Key representing a file tag group
-     */
-    private static class FileTagGroupKey extends GroupKey {
-
-        private final List<String> tagNames;
-        private final String tagNamesString;
-
-        @NbBundle.Messages({
-            "FileSearch.FileTagGroupKey.noSets=None"})
-        FileTagGroupKey(ResultFile file) {
-            tagNames = file.getTagNames();
-
-            if (tagNames.isEmpty()) {
-                tagNamesString = Bundle.FileSearch_FileTagGroupKey_noSets();
-            } else {
-                tagNamesString = String.join(",", tagNames); // NON-NLS
-            }
-        }
-
-        @Override
-        String getDisplayName() {
-            return getTagNamesString();
-        }
-
-        @Override
-        public int compareTo(GroupKey otherGroupKey) {
-            if (otherGroupKey instanceof FileTagGroupKey) {
-                FileTagGroupKey otherFileTagGroupKey = (FileTagGroupKey) otherGroupKey;
-
-                // Put the empty list at the end
-                if (getTagNames().isEmpty()) {
-                    if (otherFileTagGroupKey.getTagNames().isEmpty()) {
-                        return 0;
-                    } else {
-                        return 1;
-                    }
-                } else if (otherFileTagGroupKey.getTagNames().isEmpty()) {
-                    return -1;
-                }
-
-                return getTagNamesString().compareTo(otherFileTagGroupKey.getTagNamesString());
-            } else {
-                return compareClassNames(otherGroupKey);
-            }
-        }
-
-        @Override
-        public boolean equals(Object otherKey) {
-            if (otherKey == this) {
-                return true;
-            }
-
-            if (!(otherKey instanceof FileTagGroupKey)) {
-                return false;
-            }
-
-            FileTagGroupKey otherFileTagGroupKey = (FileTagGroupKey) otherKey;
-            return getTagNamesString().equals(otherFileTagGroupKey.getTagNamesString());
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(getTagNamesString());
-        }
-
-        /**
-         * @return the tagNames
-         */
-        List<String> getTagNames() {
-            return Collections.unmodifiableList(tagNames);
-        }
-
-        /**
-         * @return the tagNamesString
-         */
-        String getTagNamesString() {
-            return tagNamesString;
-        }
-    }
-
-    /**
-     * Default attribute used to make one group
-     */
-    static class NoGroupingAttribute extends AttributeType {
-
-        @Override
-        GroupKey getGroupKey(ResultFile file) {
-            return new NoGroupingGroupKey();
-        }
-    }
-
-    /**
-     * Dummy key for when there is no grouping. All files will have the same
-     * key.
-     */
-    private static class NoGroupingGroupKey extends GroupKey {
-
-        NoGroupingGroupKey() {
-            // Nothing to save - all files will get the same GroupKey
-        }
-
-        @NbBundle.Messages({
-            "FileSearch.NoGroupingGroupKey.allFiles=All Files"})
-        @Override
-        String getDisplayName() {
-            return Bundle.FileSearch_NoGroupingGroupKey_allFiles();
-        }
-
-        @Override
-        public int compareTo(GroupKey otherGroupKey) {
-            // As long as the other key is the same type, they are equal
-            if (otherGroupKey instanceof NoGroupingGroupKey) {
-                return 0;
-            } else {
-                return compareClassNames(otherGroupKey);
-            }
-        }
-
-        @Override
-        public boolean equals(Object otherKey) {
-            if (otherKey == this) {
-                return true;
-            }
-            // As long as the other key is the same type, they are equal
-            return otherKey instanceof NoGroupingGroupKey;
-        }
-
-        @Override
-        public int hashCode() {
-            return 0;
-        }
-    }
 
     /**
      * Enum for the attribute types that can be used for grouping.
@@ -2275,7 +1500,7 @@ class FileSearch {
         "FileSearch.GroupingAttributeType.tag.displayName=Tag",
         "FileSearch.GroupingAttributeType.object.displayName=Object Detected",
         "FileSearch.GroupingAttributeType.none.displayName=None"})
-    enum GroupingAttributeType {
+    public enum GroupingAttributeType {
         FILE_SIZE(new FileSizeAttribute(), Bundle.FileSearch_GroupingAttributeType_size_displayName()),
         FREQUENCY(new FrequencyAttribute(), Bundle.FileSearch_GroupingAttributeType_frequency_displayName()),
         KEYWORD_LIST_NAME(new KeywordListAttribute(), Bundle.FileSearch_GroupingAttributeType_keywordList_displayName()),
@@ -2285,7 +1510,7 @@ class FileSearch {
         INTERESTING_ITEM_SET(new InterestingItemAttribute(), Bundle.FileSearch_GroupingAttributeType_interestingItem_displayName()),
         FILE_TAG(new FileTagAttribute(), Bundle.FileSearch_GroupingAttributeType_tag_displayName()),
         OBJECT_DETECTED(new ObjectDetectedAttribute(), Bundle.FileSearch_GroupingAttributeType_object_displayName()),
-        NO_GROUPING(new NoGroupingAttribute(), Bundle.FileSearch_GroupingAttributeType_none_displayName());
+        NO_GROUPING(new DiscoveryKeyUtils.NoGroupingAttribute(), Bundle.FileSearch_GroupingAttributeType_none_displayName());
 
         private final AttributeType attributeType;
         private final String displayName;
@@ -2300,7 +1525,7 @@ class FileSearch {
             return displayName;
         }
 
-        AttributeType getAttributeType() {
+        public AttributeType getAttributeType() {
             return attributeType;
         }
 
@@ -2309,7 +1534,7 @@ class FileSearch {
          *
          * @return enums that can be used to group images
          */
-        static List<GroupingAttributeType> getOptionsForGrouping() {
+        public static List<GroupingAttributeType> getOptionsForGrouping() {
             return Arrays.asList(FILE_SIZE, FREQUENCY, PARENT_PATH, OBJECT_DETECTED, HASH_LIST_NAME, INTERESTING_ITEM_SET);
         }
     }
