+++ Core/src/org/sleuthkit/autopsy/discovery/search/FileSearchException.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.sleuthkit.autopsy.discovery;
+package org.sleuthkit.autopsy.discovery.search;
 
 /**
  * Exception type used for FileSearch.
