+++ Core/src/org/sleuthkit/autopsy/discovery/ui/DiscoveryUiUtils.java
@@ -18,27 +18,43 @@
  */
 package org.sleuthkit.autopsy.discovery.ui;
 
+import com.google.common.io.Files;
 import java.awt.Component;
 import java.awt.Dimension;
+import java.awt.Image;
 import java.awt.Point;
+import java.awt.image.BufferedImage;
+import java.io.IOException;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.logging.Level;
+import javax.imageio.ImageIO;
 import javax.swing.ImageIcon;
 import javax.swing.JComponent;
 import javax.swing.JOptionPane;
 import javax.swing.JScrollPane;
 import javax.swing.JTextPane;
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.FilenameUtils;
+import org.imgscalr.Scalr;
+import org.netbeans.api.progress.ProgressHandle;
+import org.opencv.core.Mat;
+import org.opencv.highgui.VideoCapture;
 import org.openide.util.ImageUtilities;
 import org.openide.util.NbBundle;
 import org.sleuthkit.autopsy.casemodule.Case;
 import org.sleuthkit.autopsy.casemodule.NoCurrentCaseException;
+import org.sleuthkit.autopsy.corelibs.ScalrWrapper;
+import org.sleuthkit.autopsy.coreutils.ImageUtils;
 import org.sleuthkit.autopsy.coreutils.Logger;
-import org.sleuthkit.autopsy.discovery.Bundle;
-import org.sleuthkit.autopsy.discovery.ResultFile;
+import static org.sleuthkit.autopsy.coreutils.VideoUtils.getVideoFileInTempDir;
+import org.sleuthkit.autopsy.datamodel.ContentUtils;
+import org.sleuthkit.autopsy.discovery.search.ResultFile;
+import org.sleuthkit.datamodel.AbstractFile;
 import org.sleuthkit.datamodel.BlackboardArtifact;
 import org.sleuthkit.datamodel.BlackboardAttribute;
 import org.sleuthkit.datamodel.DataSource;
@@ -62,6 +78,9 @@ final class DiscoveryUiUtils {
     private static final ImageIcon NOTABLE_SCORE_ICON = new ImageIcon(ImageUtilities.loadImage(RED_CIRCLE_ICON_PATH, false));
     private static final ImageIcon DELETED_ICON = new ImageIcon(ImageUtilities.loadImage(DELETE_ICON_PATH, false));
     private static final ImageIcon UNSUPPORTED_DOCUMENT_THUMBNAIL = new ImageIcon(ImageUtilities.loadImage(UNSUPPORTED_DOC_PATH, false));
+    private static final String THUMBNAIL_FORMAT = "png"; //NON-NLS
+    private static final String VIDEO_THUMBNAIL_DIR = "video-thumbnails"; //NON-NLS
+    private static final BufferedImage VIDEO_DEFAULT_IMAGE = getDefaultVideoThumbnail();
 
     @NbBundle.Messages({"# {0} - fileSize",
         "# {1} - units",
@@ -166,10 +185,11 @@ final class DiscoveryUiUtils {
      *                       false otherwise.
      * @param isDeletedLabel The label to set the icon and tooltip for.
      */
+    @NbBundle.Messages({"DiscoveryUiUtils.isDeleted.text=All instances of file are deleted."})
     static void setDeletedIcon(boolean isDeleted, javax.swing.JLabel isDeletedLabel) {
         if (isDeleted) {
             isDeletedLabel.setIcon(DELETED_ICON);
-            isDeletedLabel.setToolTipText(Bundle.ImageThumbnailPanel_isDeleted_text());
+            isDeletedLabel.setToolTipText(Bundle.DiscoveryUiUtils_isDeleted_text());
         } else {
             isDeletedLabel.setIcon(null);
             isDeletedLabel.setToolTipText(null);
@@ -244,6 +264,272 @@ final class DiscoveryUiUtils {
             logger.log(Level.WARNING, "Exception while determining which modules have been run for Discovery", ex);
         }
         dialog.validateDialog();
+    }
+
+    /**
+     * Get the video thumbnails for a file which exists in a
+     * VideoThumbnailsWrapper and update the VideoThumbnailsWrapper to include
+     * them.
+     *
+     * @param thumbnailWrapper the object which contains the file to generate
+     *                         thumbnails for.
+     *
+     */
+    @NbBundle.Messages({"# {0} - file name",
+        "DiscoveryUiUtils.genVideoThumb.progress.text=extracting temporary file {0}"})
+    static void getVideoThumbnails(VideoThumbnailsWrapper thumbnailWrapper) {
+        AbstractFile file = thumbnailWrapper.getResultFile().getFirstInstance();
+        String cacheDirectory;
+        try {
+            cacheDirectory = Case.getCurrentCaseThrows().getCacheDirectory();
+        } catch (NoCurrentCaseException ex) {
+            cacheDirectory = null;
+            logger.log(Level.WARNING, "Unable to get cache directory, video thumbnails will not be saved", ex);
+        }
+        if (cacheDirectory == null || file.getMd5Hash() == null || !Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash()).toFile().exists()) {
+            java.io.File tempFile;
+            try {
+                tempFile = getVideoFileInTempDir(file);
+            } catch (NoCurrentCaseException ex) {
+                logger.log(Level.WARNING, "Exception while getting open case.", ex); //NON-NLS
+                int[] framePositions = new int[]{
+                    0,
+                    0,
+                    0,
+                    0};
+                thumbnailWrapper.setThumbnails(createDefaultThumbnailList(VIDEO_DEFAULT_IMAGE), framePositions);
+                return;
+            }
+            if (tempFile.exists() == false || tempFile.length() < file.getSize()) {
+                ProgressHandle progress = ProgressHandle.createHandle(Bundle.DiscoveryUiUtils_genVideoThumb_progress_text(file.getName()));
+                progress.start(100);
+                try {
+                    Files.createParentDirs(tempFile);
+                    if (Thread.interrupted()) {
+                        int[] framePositions = new int[]{
+                            0,
+                            0,
+                            0,
+                            0};
+                        thumbnailWrapper.setThumbnails(createDefaultThumbnailList(VIDEO_DEFAULT_IMAGE), framePositions);
+                        return;
+                    }
+                    ContentUtils.writeToFile(file, tempFile, progress, null, true);
+                } catch (IOException ex) {
+                    logger.log(Level.WARNING, "Error extracting temporary file for " + file.getParentPath() + "/" + file.getName(), ex); //NON-NLS
+                } finally {
+                    progress.finish();
+                }
+            }
+            VideoCapture videoFile = new VideoCapture(); // will contain the video
+            BufferedImage bufferedImage = null;
+
+            try {
+                if (!videoFile.open(tempFile.toString())) {
+                    logger.log(Level.WARNING, "Error opening {0} for preview generation.", file.getParentPath() + "/" + file.getName()); //NON-NLS
+                    int[] framePositions = new int[]{
+                        0,
+                        0,
+                        0,
+                        0};
+                    thumbnailWrapper.setThumbnails(createDefaultThumbnailList(VIDEO_DEFAULT_IMAGE), framePositions);
+                    return;
+                }
+                double fps = videoFile.get(5); // gets frame per second
+                double totalFrames = videoFile.get(7); // gets total frames
+                if (fps <= 0 || totalFrames <= 0) {
+                    logger.log(Level.WARNING, "Error getting fps or total frames for {0}", file.getParentPath() + "/" + file.getName()); //NON-NLS
+                    int[] framePositions = new int[]{
+                        0,
+                        0,
+                        0,
+                        0};
+                    thumbnailWrapper.setThumbnails(createDefaultThumbnailList(VIDEO_DEFAULT_IMAGE), framePositions);
+                    return;
+                }
+                if (Thread.interrupted()) {
+                    int[] framePositions = new int[]{
+                        0,
+                        0,
+                        0,
+                        0};
+                    thumbnailWrapper.setThumbnails(createDefaultThumbnailList(VIDEO_DEFAULT_IMAGE), framePositions);
+                    return;
+                }
+
+                double duration = 1000 * (totalFrames / fps); //total milliseconds
+
+                int[] framePositions = new int[]{
+                    (int) (duration * .01),
+                    (int) (duration * .25),
+                    (int) (duration * .5),
+                    (int) (duration * .75),};
+
+                Mat imageMatrix = new Mat();
+                List<Image> videoThumbnails = new ArrayList<>();
+                if (cacheDirectory == null || file.getMd5Hash() == null) {
+                    cacheDirectory = null;
+                } else {
+                    try {
+                        FileUtils.forceMkdir(Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash()).toFile());
+                    } catch (IOException ex) {
+                        cacheDirectory = null;
+                        logger.log(Level.WARNING, "Unable to make video thumbnails directory, thumbnails will not be saved", ex);
+                    }
+                }
+                for (int i = 0; i < framePositions.length; i++) {
+                    if (!videoFile.set(0, framePositions[i])) {
+                        logger.log(Level.WARNING, "Error seeking to " + framePositions[i] + "ms in {0}", file.getParentPath() + "/" + file.getName()); //NON-NLS
+                        // If we can't set the time, continue to the next frame position and try again.
+
+                        videoThumbnails.add(VIDEO_DEFAULT_IMAGE);
+                        if (cacheDirectory != null) {
+                            try {
+                                ImageIO.write(VIDEO_DEFAULT_IMAGE, THUMBNAIL_FORMAT,
+                                        Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash(), i + "-" + framePositions[i] + "." + THUMBNAIL_FORMAT).toFile()); //NON-NLS)
+                            } catch (IOException ex) {
+                                logger.log(Level.WARNING, "Unable to save default video thumbnail for " + file.getMd5Hash() + " at frame position " + framePositions[i], ex);
+                            }
+                        }
+                        continue;
+                    }
+                    // Read the frame into the image/matrix.
+                    if (!videoFile.read(imageMatrix)) {
+                        logger.log(Level.WARNING, "Error reading frame at " + framePositions[i] + "ms from {0}", file.getParentPath() + "/" + file.getName()); //NON-NLS
+                        // If the image is bad for some reason, continue to the next frame position and try again.
+                        videoThumbnails.add(VIDEO_DEFAULT_IMAGE);
+                        if (cacheDirectory != null) {
+                            try {
+                                ImageIO.write(VIDEO_DEFAULT_IMAGE, THUMBNAIL_FORMAT,
+                                        Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash(), i + "-" + framePositions[i] + "." + THUMBNAIL_FORMAT).toFile()); //NON-NLS)
+                            } catch (IOException ex) {
+                                logger.log(Level.WARNING, "Unable to save default video thumbnail for " + file.getMd5Hash() + " at frame position " + framePositions[i], ex);
+                            }
+                        }
+
+                        continue;
+                    }
+                    // If the image is empty, return since no buffered image can be created.
+                    if (imageMatrix.empty()) {
+                        videoThumbnails.add(VIDEO_DEFAULT_IMAGE);
+                        if (cacheDirectory != null) {
+                            try {
+                                ImageIO.write(VIDEO_DEFAULT_IMAGE, THUMBNAIL_FORMAT,
+                                        Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash(), i + "-" + framePositions[i] + "." + THUMBNAIL_FORMAT).toFile()); //NON-NLS)
+                            } catch (IOException ex) {
+                                logger.log(Level.WARNING, "Unable to save default video thumbnail for " + file.getMd5Hash() + " at frame position " + framePositions[i], ex);
+                            }
+                        }
+                        continue;
+                    }
+
+                    int matrixColumns = imageMatrix.cols();
+                    int matrixRows = imageMatrix.rows();
+
+                    // Convert the matrix that contains the frame to a buffered image.
+                    if (bufferedImage == null) {
+                        bufferedImage = new BufferedImage(matrixColumns, matrixRows, BufferedImage.TYPE_3BYTE_BGR);
+                    }
+
+                    byte[] data = new byte[matrixRows * matrixColumns * (int) (imageMatrix.elemSize())];
+                    imageMatrix.get(0, 0, data); //copy the image to data
+
+                    if (imageMatrix.channels() == 3) {
+                        for (int k = 0; k < data.length; k += 3) {
+                            byte temp = data[k];
+                            data[k] = data[k + 2];
+                            data[k + 2] = temp;
+                        }
+                    }
+
+                    bufferedImage.getRaster().setDataElements(0, 0, matrixColumns, matrixRows, data);
+                    if (Thread.interrupted()) {
+                        thumbnailWrapper.setThumbnails(videoThumbnails, framePositions);
+                        try {
+                            FileUtils.forceDelete(Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash()).toFile());
+                        } catch (IOException ex) {
+                            logger.log(Level.WARNING, "Unable to delete directory for cancelled video thumbnail process", ex);
+                        }
+                        return;
+                    }
+                    BufferedImage thumbnail = ScalrWrapper.resize(bufferedImage, Scalr.Method.SPEED, Scalr.Mode.FIT_TO_HEIGHT, ImageUtils.ICON_SIZE_LARGE, ImageUtils.ICON_SIZE_MEDIUM, Scalr.OP_ANTIALIAS);
+                    //We are height limited here so it can be wider than it can be tall.Scalr maintains the aspect ratio.
+                    videoThumbnails.add(thumbnail);
+                    if (cacheDirectory != null) {
+                        try {
+                            ImageIO.write(thumbnail, THUMBNAIL_FORMAT,
+                                    Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, file.getMd5Hash(), i + "-" + framePositions[i] + "." + THUMBNAIL_FORMAT).toFile()); //NON-NLS)
+                        } catch (IOException ex) {
+                            logger.log(Level.WARNING, "Unable to save video thumbnail for " + file.getMd5Hash() + " at frame position " + framePositions[i], ex);
+                        }
+                    }
+                }
+                thumbnailWrapper.setThumbnails(videoThumbnails, framePositions);
+            } finally {
+                videoFile.release(); // close the file}
+            }
+        } else {
+            loadSavedThumbnails(cacheDirectory, thumbnailWrapper, VIDEO_DEFAULT_IMAGE);
+        }
+    }
+
+    /**
+     * Get the default image to display when a thumbnail is not available.
+     *
+     * @return The default video thumbnail.
+     */
+    private static BufferedImage getDefaultVideoThumbnail() {
+        try {
+            return ImageIO.read(ImageUtils.class
+                    .getResourceAsStream("/org/sleuthkit/autopsy/images/failedToCreateVideoThumb.png"));//NON-NLS
+        } catch (IOException ex) {
+            logger.log(Level.SEVERE, "Failed to load 'failed to create video' placeholder.", ex); //NON-NLS
+        }
+        return null;
+    }
+
+    /**
+     * Load the thumbnails that exist in the cache directory for the specified
+     * video file.
+     *
+     * @param cacheDirectory   The directory which exists for the video
+     *                         thumbnails.
+     * @param thumbnailWrapper The VideoThumbnailWrapper object which contains
+     *                         information about the file and the thumbnails
+     *                         associated with it.
+     */
+    private static void loadSavedThumbnails(String cacheDirectory, VideoThumbnailsWrapper thumbnailWrapper, BufferedImage failedVideoThumbImage) {
+        int[] framePositions = new int[4];
+        List<Image> videoThumbnails = new ArrayList<>();
+        int thumbnailNumber = 0;
+        String md5 = thumbnailWrapper.getResultFile().getFirstInstance().getMd5Hash();
+        for (String fileName : Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, md5).toFile().list()) {
+            try {
+                videoThumbnails.add(ImageIO.read(Paths.get(cacheDirectory, VIDEO_THUMBNAIL_DIR, md5, fileName).toFile()));
+            } catch (IOException ex) {
+                videoThumbnails.add(failedVideoThumbImage);
+                logger.log(Level.WARNING, "Unable to read saved video thumbnail " + fileName + " for " + md5, ex);
+            }
+            int framePos = Integer.valueOf(FilenameUtils.getBaseName(fileName).substring(2));
+            framePositions[thumbnailNumber] = framePos;
+            thumbnailNumber++;
+        }
+        thumbnailWrapper.setThumbnails(videoThumbnails, framePositions);
+    }
+
+    /**
+     * Private helper method for creating video thumbnails, for use when no
+     * thumbnails are created.
+     *
+     * @return List containing the default thumbnail.
+     */
+    private static List<Image> createDefaultThumbnailList(BufferedImage failedVideoThumbImage) {
+        List<Image> videoThumbnails = new ArrayList<>();
+        videoThumbnails.add(failedVideoThumbImage);
+        videoThumbnails.add(failedVideoThumbImage);
+        videoThumbnails.add(failedVideoThumbImage);
+        videoThumbnails.add(failedVideoThumbImage);
+        return videoThumbnails;
     }
 
     /**
