+++ Core/src/org/sleuthkit/autopsy/discovery/ui/PastOccurrencesFilterPanel.java
@@ -23,10 +23,6 @@ import javax.swing.JCheckBox;
 import javax.swing.JLabel;
 import javax.swing.JList;
 import org.sleuthkit.autopsy.centralrepository.datamodel.CentralRepository;
-import org.sleuthkit.autopsy.discovery.FileSearchData;
-import org.sleuthkit.autopsy.discovery.FileSearchData.Frequency;
-import org.sleuthkit.autopsy.discovery.SearchData.ResultType;
-import org.sleuthkit.autopsy.discovery.SearchFiltering;
 
 /**
  * Panel to allow configuration of the Past Occurrences filter.
