+++ Core/src/org/sleuthkit/autopsy/discovery/ui/SizeFilterPanel.java
@@ -18,15 +18,16 @@
  */
 package org.sleuthkit.autopsy.discovery.ui;
 
+import org.sleuthkit.autopsy.discovery.search.AbstractFilter;
 import java.util.ArrayList;
 import java.util.List;
 import javax.swing.DefaultListModel;
 import javax.swing.JCheckBox;
 import javax.swing.JLabel;
 import javax.swing.JList;
-import org.sleuthkit.autopsy.discovery.FileSearchData;
-import org.sleuthkit.autopsy.discovery.FileSearchData.FileSize;
-import org.sleuthkit.autopsy.discovery.SearchFiltering;
+import org.sleuthkit.autopsy.discovery.search.FileSearchData;
+import org.sleuthkit.autopsy.discovery.search.FileSearchData.FileSize;
+import org.sleuthkit.autopsy.discovery.search.SearchFiltering;
 
 /**
  * Panel to allow configuration of the Size Filter.
